Add Exceptions to BL

+++ BL/ReaderBL.cs
@@ -3,6 +3,7 @@ using DAL;
 using DAL.Interface;
 using Entities;
 using System.Collections.Generic;
+using System;
 
 namespace BL
 {
@@ -12,7 +13,15 @@ namespace BL
 
         public void Add(Reader reader)
         {
-            readerDao.Add(reader);
+            if (reader.isCorrectState())
+            {
+                readerDao.Add(reader);
+            }
+            else
+            {
+                throw new ArgumentException("Error, Added reader in incorrec state (empty fields)");
+            }
+            
         }
 
         public IEnumerable<Reader> GetAllReaders()
@@ -22,12 +31,26 @@ namespace BL
 
         public void Remove(int ID)
         {
-            readerDao.Remove(ID);
+            if (ID > 0)
+            {
+                readerDao.Remove(ID);
+            }
+            else
+            {
+                throw new ArgumentException("ID is negative or zero");
+            }
         }
 
         public void Update(Reader reader)
         {
-            readerDao.Update(reader);
+            if (reader.isCorrectState())
+            {
+                readerDao.Update(reader);
+            }
+            else
+            {
+                throw new ArgumentException("Error, updated reader in incorrec state (empty fields)");
+            }
         }
 
         public IEnumerable<ReaderDemo> GetAllUsers()
@@ -37,7 +60,14 @@ namespace BL
 
         public Reader GetReaderByLogin(string login)
         {
-            return readerDao.GetReaderByLogin(login);
+            if (string.IsNullOrEmpty(login))
+            {
+                throw new ArgumentException("Value empty or null -> login");
+            }
+            else
+            {
+                return readerDao.GetReaderByLogin(login);
+            }
         }
     }
 }
