+++ BL/TicketBL.cs
@@ -2,6 +2,7 @@
 using DAL;
 using DAL.Interface;
 using Entities;
+using System;
 using System.Collections.Generic;
 
 namespace BL
@@ -12,7 +13,14 @@ namespace BL
 
         public void Add(Ticket ticket)
         {
-            ticketDao.Add(ticket);
+            if (ticket.isCorrectState())
+            {
+                ticketDao.Add(ticket);
+            }
+            else
+            {
+                throw new ArgumentException("Error, Added book in incorrec state (empty fields)");
+            }
         }
 
         public IEnumerable<Ticket> GetAllTickets()
@@ -22,17 +30,39 @@ namespace BL
 
         public void Remove(int ID)
         {
-            ticketDao.Remove(ID);
+            if (ID > 0)
+            {
+                ticketDao.Remove(ID);
+            }
+            else
+            {
+                throw new ArgumentException("ID is negative or zero");
+            }
         }
 
         public void Update(Ticket ticket)
         {
-            ticketDao.Update(ticket);
+            if (ticket.isCorrectState())
+            {
+                ticketDao.Update(ticket);
+            }
+            else
+            {
+                throw new ArgumentException("Error, updated ticket in incorrec state (empty fields)");
+            }
         }
 
-        public Ticket GetTicketByID(int ticketID)
+        public Ticket GetTicketByID(int ticket_ID)
         {
-            return ticketDao.GetTicketByID(ticketID);
+            if (ticket_ID > 0)
+            {
+                return ticketDao.GetTicketByID(ticket_ID);
+            }
+            else
+            {
+                throw new ArgumentException("ticket_ID is negative or zero");
+            }
+           
         }
     }
 }
